semantic_search/requirements.txt
sentence-transformers==2.3.1
faiss-cpu==1.7.4
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2